+++ lib/synctv/client/resources/accounts/media_link.rb
@@ -1,0 +1,5 @@
+class Synctv::Client::Resources::Accounts::MediaLink < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+
+  self.prefix = "/api/v2/accounts/"
+end