+++ lib/synctv/client/resources/av_format.rb
@@ -1,0 +1,3 @@
+class Synctv::Client::Resources::AvFormat < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+end