+++ lib/synctv/client/resources/billing/credit_card.rb
@@ -1,0 +1,5 @@
+class Synctv::Client::Resources::Billing::CreditCard < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+
+  self.prefix = "/api/v2/billing/"
+end