+++ lib/synctv/client/resources/image.rb
@@ -1,0 +1,3 @@
+class Synctv::Client::Resources::Image < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+end