+++ lib/synctv/client/resources/image_format.rb
@@ -1,0 +1,5 @@
+class Synctv::Client::Resources::ImageFormat < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+  
+  self.prefix = "/api/v2/platforms/:platform_id/"
+end