+++ lib/synctv/client/resources/ip_whitelist.rb
@@ -1,0 +1,3 @@
+class Synctv::Client::Resources::IpWhitelist < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+end