+++ lib/synctv/client/resources/language.rb
@@ -1,0 +1,3 @@
+class Synctv::Client::Resources::Language < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+end