+++ lib/synctv/client/resources/manifests/smooth.rb
@@ -1,0 +1,6 @@
+class Synctv::Client::Resources::Manifests::Smooth < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+
+  self.prefix       = "/api/v2/media/:medium_id/"
+  self.element_name = "smooth_manifest"
+end