+++ lib/synctv/client/resources/media/view.rb
@@ -1,0 +1,5 @@
+class Synctv::Client::Resources::Media::View < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+
+  self.prefix       = "/api/v2/media/"
+end