+++ lib/synctv/client/resources/subtitle.rb
@@ -1,0 +1,3 @@
+class Synctv::Client::Resources::Subtitle < Synctv::Client::Resource
+  include Synctv::Client::Scopes
+end